// src/drivers/dto/driver.input.dto.ts
import {VehicleFeaturesType} from "../types/drivers";

export type DriverInputDto = {
    name: string;
    phoneNumber: string;
    email: string;
    vehicleMake: string;
    vehicleModel: string;
    vehicleYear: number;
    vehicleLicensePlate: string;
    vehicleDescription: string | null;
    vehicleFeatures: VehicleFeaturesType[];
};