// __tests__/drivers-body-validation.e2e.spec.ts
import request from 'supertest';
import express from 'express';
import {DriverInputDto} from "../src/drivers/dto/driver.input.dto";
import {setupApp} from "../src/setup-app";
import {VehicleFeaturesType} from "../src/drivers/types/drivers";
import {HttpStatus} from "../src/core/types/http-statuses";

describe('Driver API body validation check', () => {
    const app = express();
    setupApp(app);

    const correctTestDriverData: DriverInputDto = {
        name: 'Valentin',
        phoneNumber: '123-456-7890',
        email: 'valentin@example.com',
        vehicleMake: 'BMW',
        vehicleModel: 'X5',
        vehicleYear: 2021,
        vehicleLicensePlate: 'ABC-123',
        vehicleDescription: 'Some description',
        vehicleFeatures: [VehicleFeaturesType.ChildSeat],
    };

    beforeAll(async () => {
        await request(app).delete('/testing/all-data').expect(HttpStatus.NoContent);
    });

    it(`should not create driver when incorrect body passed; POST /drivers`, async () => {
        const invalidDataSet1 = await request(app)
            .post('/drivers')
            .send({
                ...correctTestDriverData,
                name: '   ',
                phoneNumber: '    ',
                email: 'invalid email',
                vehicleMake: '',
            })
            .expect(HttpStatus.BadRequest);

        expect(invalidDataSet1.body.errorMessages).toHaveLength(4);

        const invalidDataSet2 = await request(app)
            .post('/drivers')
            .send({
                ...correctTestDriverData,
                phoneNumber: '',
                vehicleModel: '',
                vehicleYear: 'year',
                vehicleLicensePlate: '',
            })
            .expect(HttpStatus.BadRequest);

        expect(invalidDataSet2.body.errorMessages).toHaveLength(4);

        const invalidDataSet3 = await request(app)
            .post('/drivers')
            .send({
                ...correctTestDriverData,
                name: 'A',
            })
            .expect(HttpStatus.BadRequest);

        expect(invalidDataSet3.body.errorMessages).toHaveLength(1);


        const driverListResponse = await request(app).get('/drivers');
        expect(driverListResponse.body).toHaveLength(0);
    });
});